\name{ClockstaRG-package}
\alias{ClockstaRG-package}
\alias{ClockstaRG}
\docType{package}
\title{
ClockstaR-G: ClockstaR for genomic data sets
}
\description{
ClockstaR-G is an implementation of ClockstaR for large data sets. It uses multicore packages to speed computation. It also includes a modification of the initial ClockstaR algorithm.
}
\details{
\tabular{ll}{
Package: \tab ClockstarRG\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-12-25\cr
License: \tab GNU\cr
}
Please see examples in github
}
\author{
Sebastian Duchene
Maintainer: Sebastian Duchene <sebastian.duchene@sydney.edu.au>
}
\references{
Please see original paper. 
}
\keyword{ ClockstaR, genome pacemakers }
\seealso{
ClockstaR
}
